Replace unsigned int with size_t where value indicates a count.

+++ include/hashtable.h
@@ -45,18 +45,18 @@ hashtable*	hashtable_new(dict_compare_func cmp_func, dict_hash_func hash_func,
 dict*		hashtable_dict_new(dict_compare_func cmp_func,
 							   dict_hash_func hash_func,
 							   dict_delete_func del_func, unsigned size);
-unsigned	hashtable_free(hashtable *table);
+size_t		hashtable_free(hashtable *table);
 
 int			hashtable_insert(hashtable *table, void *key, void *datum,
 							 int overwrite);
 int			hashtable_probe(hashtable *table, void *key, void **datum);
 void*		hashtable_search(hashtable *table, const void *key);
 int			hashtable_remove(hashtable *table, const void *key);
-unsigned	hashtable_clear(hashtable *table);
-unsigned	hashtable_traverse(hashtable *table, dict_visit_func visit);
-unsigned	hashtable_count(const hashtable *table);
-unsigned	hashtable_size(const hashtable *table);
-unsigned	hashtable_slots_used(const hashtable *table);
+size_t		hashtable_clear(hashtable *table);
+size_t		hashtable_traverse(hashtable *table, dict_visit_func visit);
+size_t		hashtable_count(const hashtable *table);
+size_t		hashtable_size(const hashtable *table);
+size_t		hashtable_slots_used(const hashtable *table);
 int			hashtable_resize(hashtable *table, unsigned size);
 
 typedef struct hashtable_itor
@@ -71,8 +71,8 @@ int			hashtable_itor_valid(const hashtable_itor *itor);
 void		hashtable_itor_invalidate(hashtable_itor *itor);
 int			hashtable_itor_next(hashtable_itor *itor);
 int			hashtable_itor_prev(hashtable_itor *itor);
-int			hashtable_itor_nextn(hashtable_itor *itor, unsigned count);
-int			hashtable_itor_prevn(hashtable_itor *itor, unsigned count);
+int			hashtable_itor_nextn(hashtable_itor *itor, size_t count);
+int			hashtable_itor_prevn(hashtable_itor *itor, size_t count);
 int			hashtable_itor_first(hashtable_itor *itor);
 int			hashtable_itor_last(hashtable_itor *itor);
 int			hashtable_itor_search(hashtable_itor *itor, const void *key);
